📝 feat. CardBage폴더명 및 경로변경
