+++ src/components/listPage/cardBadge/CardBaged.style.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+import { theme } from '../../../style/theme'
+
+export const Div = styled.div`
+  display: flex;
+  margin: -3px 8px 20px;
+  width: 100%;
+  box-sizing: border-box;
+  font-size: inherit;
+`
